Added files for the emotion classifier along with the trained model.
This commit is made to add the emotional classifier files which predict the emotions for the text generated by the LLM and which is later used to generate the payloads to set up the emotions. If training the model from scratch, be sure to change the paths for the dataset along with the emotion mappings.